Revert "Fix footnote"
