+++ backend/routes/documents.js
@@ -289,7 +289,7 @@ router.post("/", authenticate, upload.single("file"), async (req, res) => {
           canWrite: true,
         },
       });
-      const version = await tx.version.create({
+      await tx.version.create({
         data: {
           content: { ops: delta },
           docId: doc.id,
@@ -297,12 +297,6 @@ router.post("/", authenticate, upload.single("file"), async (req, res) => {
         },
       });
 
-      await tx.doc.update({
-        where: { id: doc.id },
-        data: {
-          currentVersionId: version.id,
-        },
-      });
       return doc;
     });
 
@@ -1071,7 +1065,6 @@ router.patch("/:id", authenticate, async (req, res) => {
 router.patch("/:id/content", authenticate, async (req, res) => {
   const { docs_prosemirror_delta } = req.body;
   try {
-    // First, get the document with its current version
     const document = await prisma.doc.findUnique({
       where: { id: req.params.id },
       select: {
@@ -1079,20 +1072,10 @@ router.patch("/:id/content", authenticate, async (req, res) => {
         id: true,
         docs_prosemirror_delta: true,
         rootProjectId: true,
-        currentVersionId: true,
-        currentVersion: {
-          select: {
-            id: true,
-            content: true,
-            label: true,
-          },
-        },
       },
     });
-
     if (!document) return res.status(404).json({ error: "Document not found" });
 
-    // Check permissions
     if (document.ownerId !== req.user.id) {
       const permission = await prisma.permission.findFirst({
         where: {
@@ -1104,69 +1087,92 @@ router.patch("/:id/content", authenticate, async (req, res) => {
       if (!permission) return res.status(403).json({ error: "No edit access" });
     }
 
-    await prisma.doc.update({
+    // Approach 1: If we need to merge with existing Y.doc state
+    const ydoc = new Y.Doc({ gc: true });
+
+    // If the document already has Y.doc state, we should first apply that
+    // if (document.docs_y_doc_state) {
+    //   Y.applyUpdateV2(ydoc, document.docs_y_doc_state);
+    // }
+
+    // Get the shared text type from the Y.doc
+    const ytext = ydoc.getText(req.params.id);
+
+    // Convert ProseMirror delta to Y.js compatible operations
+    // This is the critical part that was missing
+    if (docs_prosemirror_delta) {
+      // Process the delta operations to make them compatible with Y.js
+      // You may need a custom conversion function depending on your delta format
+      const yDelta = convertProseMirrorDeltaToYDelta(docs_prosemirror_delta);
+      ytext.applyDelta(yDelta);
+    }
+
+    // Encode the updated Y.doc state
+
+    const docs_y_doc_state = Y.encodeStateAsUpdateV2(ydoc);
+
+    // Validate the state isn't too small/empty
+    // if (docs_y_doc_state.length < 100) {
+    //   console.log('Y.js state is too small, skipping update');
+    //   return res.status(400).json({ error: "Invalid document state" });
+    // }
+
+    // Uncomment this to actually update the database
+    const updatedDocument = await prisma.doc.update({
       where: { id: document.id },
       data: {
         docs_prosemirror_delta,
       },
+      select: {
+        id: true,
+      },
     });
-
-    let currentVersionId = document.currentVersionId;
-    // First, check if this is a system-generated version (initial auto-save)
-    const currentVersion = await prisma.version.findUnique({
-      where: { id: currentVersionId },
-      select: { userId: true, label: true, content: true },
-    });
-
-    let newVersion;
-    // Prevent updating system-generated versions (initial auto-save), instead create a new version
-    if (!currentVersion?.userId) {
-      // Compare new content with current version content
-      const currentContent = JSON.stringify(currentVersion.content?.ops);
-      const newContent = JSON.stringify(docs_prosemirror_delta );
-      if (currentContent !== newContent) {
-        newVersion = await prisma.version.create({
-          data: {
-            content: {ops:docs_prosemirror_delta},
-            docId: document.id,
-            label: "Edited Initial Auto-save",
-            userId: req.user.id,
-          },
-        });
-
-        // Update the document's currentVersionId to point to the new version
-        await prisma.doc.update({
-          where: { id: document.id },
-          data: {
-            currentVersionId: newVersion.id,
-          },
-        });
-
-        currentVersionId = newVersion.id;
-      } else {
-        console.log("Content is the same, skipping version creation");
-      }
-      // If content is the same, don't create a new version - keep using the existing one
-    } else {
-      // Update the existing user version
-      await prisma.version.update({
-        where: { id: currentVersionId },
-        data: {
-          content: { ops: docs_prosemirror_delta || {} },
-        },
-      });
-    }
-
     res.json({
       success: true,
-      currentVersionId: currentVersionId,
-      message: "Version content updated successfully",
     });
   } catch (error) {
-    console.error("Error updating version content:", error);
-    res.status(500).json({ error: "Error updating version content" });
+    console.error("Error updating document content:", error);
+    res.status(500).json({ error: "Error updating document content" });
   }
 });
+
+/**
+ * Convert ProseMirror delta format to Y.js delta format
+ * You'll need to customize this based on your specific delta format
+ */
+function convertProseMirrorDeltaToYDelta(prosemirrorDelta) {
+  // This is a simplified example - your actual conversion will depend on
+  // the specific structure of your ProseMirror delta
+  const yDelta = [];
+
+  // If prosemirrorDelta is an array of operations
+  if (Array.isArray(prosemirrorDelta)) {
+    for (const op of prosemirrorDelta) {
+      if (op.insert) {
+        // Convert insert operations
+        yDelta.push({ insert: op.insert });
+      } else if (op.delete) {
+        // Convert delete operations
+        yDelta.push({ delete: op.delete });
+      } else if (op.retain) {
+        // Convert retain operations
+        yDelta.push({ retain: op.retain });
+      }
+      // Add attributes if they exist
+      if (op.attributes) {
+        const lastOp = yDelta[yDelta.length - 1];
+        if (lastOp) {
+          lastOp.attributes = op.attributes;
+        }
+      }
+    }
+  } else if (typeof prosemirrorDelta === "object") {
+    // If it's a single operation object
+    return [prosemirrorDelta];
+  }
+
+  return yDelta;
+}
 
 // Alternative approach: If you're using prosemirror-y-binding or a similar library
 // You might want to use their built-in conversion functions instead
@@ -1296,6 +1302,8 @@ router.post("/generate-translation", authenticate, async (req, res) => {
       process.env.SERVER_URL || `http://localhost:${process.env.PORT || 9000}`;
     const webhookUrl = `${serverUrl}/documents/translation-webhook/${translationId}`;
 
+    console.log(`Setting up webhook URL: ${webhookUrl}`);
+
     // Prepare translation request data
     const translationData = {
       api_key: apiKey,
@@ -1422,7 +1430,9 @@ router.post("/translation-webhook/:id", async (req, res) => {
       });
       return updated;
     });
-
+    console.log(
+      `Translation webhook received and processed for document ${document_id}`
+    );
     res.status(200).json({ success: true, data });
   } catch (error) {
     console.error("Error processing translation webhook:", error);
