+++ backend/routes/translate.js
@@ -183,206 +183,111 @@ router.post("/", authenticate, async (req, res) => {
           for (const line of lines) {
             if (!line.trim()) continue;
 
-            // Extract event data from SSE format
-            let eventData = "";
-            if (line.trim().startsWith("data: ")) {
-              eventData = line.trim().substring(6).trim();
-            } else if (line.trim().startsWith("data:")) {
-              eventData = line.trim().substring(5).trim();
+            // Parse Server-Sent Events format from external API
+            let eventData = line;
+
+            // Handle SSE format: "data: ..."
+            if (line.startsWith("data: ")) {
+              eventData = line.substring(6); // Remove "data: " prefix
+
+              // Handle nested SSE format: "data: data: {...}"
+              if (eventData.startsWith("data: ")) {
+                eventData = eventData.substring(6); // Remove second "data: " prefix
+              }
+
+              // Remove carriage return
+              eventData = eventData.replace(/\r$/, "");
             }
 
-            if (!eventData) continue;
+            // Skip empty data
+            if (!eventData.trim()) continue;
 
-            // Handle multiple JSON objects or fragments in eventData
-            if (eventData.startsWith("{") && eventData.endsWith("}")) {
-              try {
-                // Single complete JSON object
-                const data = JSON.parse(eventData);
+            try {
+              // Parse the JSON data
+              const data = JSON.parse(eventData);
 
-                // Transform external API events to our structured format
-                if (data.type === "initialization") {
-                  // External API initialization - we already sent ours
-                  console.log("External API initialized");
-                } else if (data.type === "planning") {
-                  // External API planning - we already sent ours
-                  console.log("External API planned batches");
-                } else if (data.type === "batch_start") {
-                  sendEvent({
-                    type: "batch_start",
-                    batch_number: data.batch_number || currentBatch,
-                    progress_percent:
-                      data.progress_percent ||
-                      ((currentBatch - 1) / totalBatches) * 100,
-                    timestamp: new Date().toISOString(),
-                    message: `Processing batch ${
-                      data.batch_number || currentBatch
-                    }...`,
-                  });
-                } else if (data.type === "translation_start") {
-                  sendEvent({
-                    type: "translation_start",
-                    timestamp: new Date().toISOString(),
-                    message: "Translating...",
-                  });
-                } else if (data.type === "text_completed") {
-                  completedTexts++;
-                  const progressPercent = (completedTexts / texts.length) * 100;
+              // Transform external API events to our structured format
+              if (data.type === "initialization") {
+                // External API initialization - we already sent ours
+                console.log("External API initialized");
+              } else if (data.type === "planning") {
+                // External API planning - we already sent ours
+                console.log("External API planned batches");
+              } else if (data.type === "batch_start") {
+                sendEvent({
+                  type: "batch_start",
+                  batch_number: data.batch_number || currentBatch,
+                  progress_percent:
+                    data.progress_percent ||
+                    ((currentBatch - 1) / totalBatches) * 100,
+                  timestamp: new Date().toISOString(),
+                  message: `Processing batch ${
+                    data.batch_number || currentBatch
+                  }...`,
+                });
+              } else if (data.type === "translation_start") {
+                sendEvent({
+                  type: "translation_start",
+                  timestamp: new Date().toISOString(),
+                  message: "Translating...",
+                });
+              } else if (data.type === "text_completed") {
+                completedTexts++;
+                const progressPercent = (completedTexts / texts.length) * 100;
 
-                  sendEvent({
-                    type: "text_completed",
-                    text_number: completedTexts,
-                    total_texts: texts.length,
-                    progress_percent: progressPercent,
-                    translation_preview: data.translation_preview,
-                    timestamp: new Date().toISOString(),
-                    message: `Completed ${completedTexts}/${texts.length} texts`,
-                  });
-                } else if (data.type === "batch_completed") {
-                  const cumulativeProgress =
-                    (data.batch_number / totalBatches) * 100;
+                sendEvent({
+                  type: "text_completed",
+                  text_number: completedTexts,
+                  total_texts: texts.length,
+                  progress_percent: progressPercent,
+                  translation_preview: data.translation_preview,
+                  timestamp: new Date().toISOString(),
+                  message: `Completed ${completedTexts}/${texts.length} texts`,
+                });
+              } else if (data.type === "batch_completed") {
+                const cumulativeProgress =
+                  (data.batch_number / totalBatches) * 100;
 
-                  sendEvent({
-                    type: "batch_completed",
-                    batch_number: data.batch_number || currentBatch,
-                    batch_id: data.batch_id || `batch_${currentBatch}`,
-                    batch_results: data.batch_results || [],
-                    cumulative_progress: cumulativeProgress,
-                    processing_time: data.processing_time || "0.0",
-                    timestamp: new Date().toISOString(),
-                    message: `Batch ${
-                      data.batch_number || currentBatch
-                    } completed in ${data.processing_time || "0.0"}s`,
-                  });
+                sendEvent({
+                  type: "batch_completed",
+                  batch_number: data.batch_number || currentBatch,
+                  batch_id: data.batch_id || `batch_${currentBatch}`,
+                  batch_results: data.batch_results || [],
+                  cumulative_progress: cumulativeProgress,
+                  processing_time: data.processing_time || "0.0",
+                  timestamp: new Date().toISOString(),
+                  message: `Batch ${
+                    data.batch_number || currentBatch
+                  } completed in ${data.processing_time || "0.0"}s`,
+                });
 
-                  currentBatch = (data.batch_number || currentBatch) + 1;
-                } else if (data.type === "completion") {
-                  sendEvent({
-                    type: "completion",
-                    total_completed:
-                      data.successful_translations || completedTexts,
-                    total_texts: data.total_texts || texts.length,
-                    timestamp: new Date().toISOString(),
-                    message: "Translation completed!",
-                  });
-                } else if (data.type === "error") {
-                  sendEvent({
-                    type: "error",
-                    error: data.error,
-                    details: data.details,
-                    timestamp: new Date().toISOString(),
-                  });
-                } else {
-                  // Forward any other structured data as-is
-                  sendEvent({
-                    ...data,
-                    timestamp: data.timestamp || new Date().toISOString(),
-                  });
-                }
-              } catch (parseError) {
-                console.error(
-                  "Error parsing complete JSON event:",
-                  parseError,
-                  "Raw eventData:",
-                  eventData
-                );
+                currentBatch = (data.batch_number || currentBatch) + 1;
+              } else if (data.type === "completion") {
+                sendEvent({
+                  type: "completion",
+                  total_completed:
+                    data.successful_translations || completedTexts,
+                  total_texts: data.total_texts || texts.length,
+                  timestamp: new Date().toISOString(),
+                  message: "Translation completed!",
+                });
+              } else if (data.type === "error") {
+                sendEvent({
+                  type: "error",
+                  error: data.error,
+                  details: data.details,
+                  timestamp: new Date().toISOString(),
+                });
+              } else {
+                // Forward any other structured data as-is
+                sendEvent({
+                  ...data,
+                  timestamp: data.timestamp || new Date().toISOString(),
+                });
               }
-            } else {
-              // Handle partial or malformed JSON using a more robust approach
-              try {
-                // Try to fix common SSE formatting issues
-                let cleanedData = eventData;
-
-                // Remove duplicate "data: " prefixes that sometimes occur
-                if (cleanedData.includes("data: ")) {
-                  cleanedData = cleanedData.replace(/data:\s*/g, "");
-                }
-
-                // Try to parse after cleaning
-                if (cleanedData.startsWith("{")) {
-                  const data = JSON.parse(cleanedData);
-
-                  // Transform external API events to our structured format
-                  if (data.type === "initialization") {
-                    // External API initialization - we already sent ours
-                    console.log("External API initialized");
-                  } else if (data.type === "planning") {
-                    // External API planning - we already sent ours
-                    console.log("External API planned batches");
-                  } else if (data.type === "batch_start") {
-                    sendEvent({
-                      type: "batch_start",
-                      batch_number: data.batch_number || currentBatch,
-                      progress_percent:
-                        data.progress_percent ||
-                        ((currentBatch - 1) / totalBatches) * 100,
-                      timestamp: new Date().toISOString(),
-                      message: `Processing batch ${
-                        data.batch_number || currentBatch
-                      }...`,
-                    });
-                  } else if (data.type === "translation_start") {
-                    sendEvent({
-                      type: "translation_start",
-                      timestamp: new Date().toISOString(),
-                      message: data.message || "Starting translation...",
-                    });
-                  } else if (data.type === "batch_completed") {
-                    currentBatch++;
-                    completedTexts += data.batch_results?.length || 0;
-
-                    sendEvent({
-                      type: "batch_completed",
-                      batch_number: data.batch_number || currentBatch - 1,
-                      batch_id: data.batch_id,
-                      batch_results: data.batch_results || [],
-                      cumulative_progress: Math.round(
-                        (completedTexts / texts.length) * 100
-                      ),
-                      processing_time:
-                        data.processing_time || "Processing time unknown",
-                      timestamp: new Date().toISOString(),
-                      message: `Batch ${
-                        data.batch_number || currentBatch - 1
-                      } completed`,
-                    });
-                  } else if (data.type === "completion") {
-                    sendEvent({
-                      type: "completion",
-                      total_completed: data.total_completed || completedTexts,
-                      total_texts: texts.length,
-                      processing_time:
-                        data.processing_time || "Processing time unknown",
-                      timestamp: new Date().toISOString(),
-                      message: "Translation completed successfully!",
-                    });
-                  } else if (data.type === "error") {
-                    sendEvent({
-                      type: "error",
-                      error: data.error,
-                      details: data.details,
-                      timestamp: new Date().toISOString(),
-                    });
-                  } else {
-                    // Forward any other structured data as-is
-                    sendEvent({
-                      ...data,
-                      timestamp: data.timestamp || new Date().toISOString(),
-                    });
-                  }
-                } else {
-                  console.warn("Skipping non-JSON event data:", eventData);
-                }
-              } catch (parseError) {
-                console.error(
-                  "Error parsing SSE event:",
-                  parseError,
-                  "Raw line:",
-                  line,
-                  "Extracted eventData:",
-                  eventData
-                );
-                // Continue processing other lines
-              }
+            } catch (parseError) {
+              // If not JSON, log but don't forward raw content
+              console.log("Non-JSON line from external API:", line);
             }
           }
         }
