+++ frontend/src/api/translate.ts
@@ -1,4 +1,4 @@
-import { getHeaders } from "./utils";
+import { getHeaders, getAuthToken } from "./utils";
 
 const server_url = import.meta.env.VITE_SERVER_URL;
 
@@ -156,7 +156,7 @@ export interface BatchCompletedEvent extends StreamEvent {
   batch_results: Array<{
     original_text: string;
     translated_text: string;
-    metadata?: Record<string, unknown>;
+    metadata?: any;
   }>;
   cumulative_progress: number;
   processing_time: string;
@@ -249,84 +249,37 @@ export const consumeTranslationStream = async (
 
         if (!eventData) continue;
 
-        // Handle multiple JSON objects or fragments in eventData
-        if (eventData.startsWith("{") && eventData.endsWith("}")) {
-          try {
-            // Single complete JSON object
-            const parsedEvent = JSON.parse(eventData) as TranslationStreamEvent;
+        try {
+          const parsedEvent = JSON.parse(eventData) as TranslationStreamEvent;
 
-            // Handle error events immediately
-            if (parsedEvent.type === "error") {
-              const error = new Error(
-                `Translation error: ${parsedEvent.error}`
-              );
-              onError?.(error);
-              return; // Stop processing on error
-            }
-
-            // Call the event handler with the parsed event
-            onEvent(parsedEvent);
-          } catch (parseError) {
-            console.warn(
-              "Failed to parse complete JSON event:",
-              parseError,
-              "Raw eventData:",
-              eventData
-            );
+          // Handle error events immediately
+          if (parsedEvent.type === "error") {
+            const error = new Error(`Translation error: ${parsedEvent.error}`);
+            onError?.(error);
+            return; // Stop processing on error
           }
-        } else {
-          // Handle partial or malformed JSON using a more robust approach
-          try {
-            // Try to fix common SSE formatting issues
-            let cleanedData = eventData;
 
-            // Remove duplicate "data: " prefixes that sometimes occur
-            if (cleanedData.includes("data: ")) {
-              cleanedData = cleanedData.replace(/data:\s*/g, "");
-            }
+          // Call the event handler with the parsed event
+          onEvent(parsedEvent);
+        } catch (parseError) {
+          console.warn(
+            "Failed to parse streaming event:",
+            parseError,
+            "Raw line:",
+            line
+          );
 
-            // Try to parse after cleaning
-            if (cleanedData.startsWith("{")) {
-              const parsedEvent = JSON.parse(
-                cleanedData
-              ) as TranslationStreamEvent;
-
-              // Handle error events immediately
-              if (parsedEvent.type === "error") {
-                const error = new Error(
-                  `Translation error: ${parsedEvent.error}`
-                );
-                onError?.(error);
-                return; // Stop processing on error
-              }
-
-              // Call the event handler with the parsed event
-              onEvent(parsedEvent);
-            } else {
-              console.warn("Skipping non-JSON event data:", eventData);
-            }
-          } catch (parseError) {
-            console.warn(
-              "Failed to parse streaming event:",
-              parseError,
-              "Raw line:",
-              line,
-              "Extracted eventData:",
-              eventData
+          // Check for authentication errors in raw text
+          if (
+            line.toLowerCase().includes("authentication") ||
+            line.toLowerCase().includes("unauthorized") ||
+            line.toLowerCase().includes("401")
+          ) {
+            const error = new Error(
+              "Authentication error during translation. Please log in again."
             );
-
-            // Check for authentication errors in raw text
-            if (
-              line.toLowerCase().includes("authentication") ||
-              line.toLowerCase().includes("unauthorized") ||
-              line.toLowerCase().includes("401")
-            ) {
-              const error = new Error(
-                "Authentication error during translation. Please log in again."
-              );
-              onError?.(error);
-              return;
-            }
+            onError?.(error);
+            return;
           }
         }
       }
