+++ frontend/src/components/Toolbar/QuillVersionControls.tsx
@@ -10,7 +10,6 @@ const VersionControls = ({
   openHistory: boolean;
   setShowVersionDiff: (show: boolean) => void;
 }) => {
-
   const {  
     createNamedSnapshot,
     isCreatingVersion,
