+++ frontend/src/components/Toolbar/VersionList.tsx
@@ -1,11 +1,12 @@
 import { useQuillVersion } from "@/contexts/VersionContext";
-import { MdDelete } from "react-icons/md";
+import { MdDelete, MdSave } from "react-icons/md";
 import { SiTicktick } from "react-icons/si";
 import { FaSpinner } from "react-icons/fa";
 import formatTimeAgo from "@/lib/formatTimeAgo";
 import { useState } from "react";
 import { ConfirmationModal } from "@/components/ui/ConfirmationModal";
 import { createPortal } from "react-dom";
+
 
 // Use the Version type from context
 
@@ -45,8 +46,8 @@ function VersionList({ handleViewAll }: { handleViewAll: () => void }) {
   const handleDeleteConfirm = async () => {
     if (!deleteModalState.versionId) return;
 
-    setDeleteModalState((prev) => ({ ...prev, isDeleting: true, error: null }));
-
+    setDeleteModalState(prev => ({ ...prev, isDeleting: true, error: null }));
+    
     try {
       await deleteVersion(deleteModalState.versionId);
       setDeleteModalState({
@@ -58,7 +59,7 @@ function VersionList({ handleViewAll }: { handleViewAll: () => void }) {
       });
     } catch (error) {
       console.error("Failed to delete version:", error);
-      setDeleteModalState((prev) => ({
+      setDeleteModalState(prev => ({
         ...prev,
         isDeleting: false,
         error: "Failed to delete version. Please try again.",
@@ -93,14 +94,11 @@ function VersionList({ handleViewAll }: { handleViewAll: () => void }) {
         ) : (
           <div className="max-h-60 overflow-y-auto border">
             {versions.map((version: any) => (
-              <EachVersion
-                key={version.id}
-                version={version}
+              <EachVersion 
+                key={version.id} 
+                version={version} 
                 onDeleteClick={handleDeleteClick}
-                isDeleting={
-                  deleteModalState.isDeleting &&
-                  deleteModalState.versionId === version.id
-                }
+                isDeleting={deleteModalState.isDeleting && deleteModalState.versionId === version.id}
               />
             ))}
           </div>
@@ -111,12 +109,10 @@ function VersionList({ handleViewAll }: { handleViewAll: () => void }) {
         <ConfirmationModal
           open={deleteModalState.isOpen}
           onClose={handleCloseModal}
-          onConfirm={
-            deleteModalState.error ? handleCloseModal : handleDeleteConfirm
-          }
+          onConfirm={deleteModalState.error ? handleCloseModal : handleDeleteConfirm}
           title={deleteModalState.error ? "Error" : "Delete Version"}
           message={
-            deleteModalState.error
+            deleteModalState.error 
               ? deleteModalState.error
               : `Are you sure you want to delete the version "${deleteModalState.versionLabel}"? This action cannot be undone.`
           }
@@ -130,24 +126,43 @@ function VersionList({ handleViewAll }: { handleViewAll: () => void }) {
   );
 }
 
+
 function EachVersion({ version, onDeleteClick, isDeleting }: EachVersionProps) {
-  const {
-    currentVersionId,
-    loadVersion,
-    isLoadingVersion,
+  const { 
+    currentVersionId, 
+    loadVersion, 
+    isLoadingVersion, 
     loadingVersionId,
     versions,
+    updateCurrentVersion,
   } = useQuillVersion();
-
+  
+  const [isSaving, setIsSaving] = useState(false);
+  const [saveError, setSaveError] = useState<string | null>(null);
+  
   const isLoading = isLoadingVersion && loadingVersionId === version.id;
   const isCurrentVersion = version.id === currentVersionId;
   const isLatestVersion = versions[0]?.id === version.id;
   const isSystemVersion = version.user === null;
-  const canDelete =
-    isLatestVersion &&
-    !isLoading &&
-    !isDeleting &&
-    !isSystemVersion;
+  const canDelete = isLatestVersion && !isLoading && !isSaving && !isDeleting && !isSystemVersion;
+
+  const handleSave = async (e: React.MouseEvent) => {
+    e.stopPropagation();
+    if (isSaving) return;
+    
+    setIsSaving(true);
+    setSaveError(null);
+    
+    try {
+      await updateCurrentVersion();
+      console.log("Successfully saved current version:", version.label);
+    } catch (error) {
+      console.error("Failed to save version:", error);
+      setSaveError("Failed to save version");
+    } finally {
+      setIsSaving(false);
+    }
+  };
 
   const handleLoad = (e: React.MouseEvent) => {
     e.stopPropagation();
@@ -163,26 +178,29 @@ function EachVersion({ version, onDeleteClick, isDeleting }: EachVersionProps) {
   };
 
   return (
-    <div
-      className={`px-2 py-2 border-b hover:bg-gray-100 ${
-        isCurrentVersion ? "bg-blue-100" : ""
-      }`}
-    >
+    <div className={`px-2 py-2 border-b hover:bg-gray-100 ${isCurrentVersion ? "bg-blue-100" : ""}`}>
       <div className="flex justify-between items-center">
-        <div
-          className={`flex items-center gap-1 ${
-            isCurrentVersion ? "font-semibold text-blue-600" : ""
-          }`}
-        >
+        <div className={`flex items-center gap-1 ${isCurrentVersion ? "font-semibold text-blue-600" : ""}`}>
           {version.label}
         </div>
         <div className="flex gap-2 justify-end">
-          {/* First button slot: Load (non-current) or invisible placeholder (current) */}
+          {/* First button slot: Save (current non-system) or Load (non-current) */}
+          {isCurrentVersion && !isSystemVersion && (
+            <button
+              onClick={handleSave}
+              disabled={isLoading || isSaving}
+              className="px-2 py-1 rounded text-sm bg-green-100 hover:bg-green-200 text-green-700"
+            >
+              {(isLoading || isSaving) ? (
+                <FaSpinner className="animate-spin" />
+              ) : (
+                <MdSave />
+              )}
+            </button>
+          )}
           
-           {isCurrentVersion ? 
-           <div className="px-2 py-1 w-8">
-           </div> : <button
-
+          {(!isCurrentVersion || isSystemVersion) && (
+            <button
               onClick={handleLoad}
               disabled={isLoading}
               className="px-2 py-1 rounded text-sm bg-gray-200 hover:bg-gray-300"
@@ -193,9 +211,8 @@ function EachVersion({ version, onDeleteClick, isDeleting }: EachVersionProps) {
                 <SiTicktick />
               )}
             </button>
-            }
+          )}
           
-
           {/* Second button slot: Delete (non-system) or invisible placeholder (system) */}
           {!isSystemVersion ? (
             <button
@@ -205,11 +222,7 @@ function EachVersion({ version, onDeleteClick, isDeleting }: EachVersionProps) {
                 canDelete ? "bg-red-100 hover:bg-red-200" : "invisible"
               }`}
             >
-              {isDeleting ? (
-                <FaSpinner className="animate-spin" />
-              ) : (
-                <MdDelete />
-              )}
+              {isDeleting ? <FaSpinner className="animate-spin" /> : <MdDelete />}
             </button>
           ) : (
             <div className="px-2 py-1 w-8"></div>
@@ -217,13 +230,11 @@ function EachVersion({ version, onDeleteClick, isDeleting }: EachVersionProps) {
         </div>
       </div>
 
-      <div
-        className={`text-xs mt-1 ${
-          isCurrentVersion ? "text-blue-600" : "text-gray-500"
-        }`}
-      >
-        {version?.user?.username || version?.user?.name || "System"} •{" "}
-        {formatTimeAgo(version.createdAt)}
+      <div className={`text-xs mt-1 ${isCurrentVersion ? "text-blue-600" : "text-gray-500"}`}>
+        {version?.user?.username || version?.user?.name || "System"} • {formatTimeAgo(version.timestamp)}
+        {saveError && isCurrentVersion && (
+          <div className="text-red-500 mt-1">{saveError}</div>
+        )}
       </div>
     </div>
   );
